(feat) Add service block to goreleaser brew config
This block allows for dagu to be run as a service under homebrew
and it then allows the following commands:

```
brew services status dagu
brew services start dagu
brew services stop dagu
```

When the brew install happens, there is also a prompt to the user
in case they want to enable the service or run directly.

The benefit here is that it enables dagu to be run as a background
service automatically on startup and controlled via a centralized
system.